+++ Changelog.md
@@ -1,16 +1,18 @@
 # Changelog
  
+## 1.4.1a
+* Restriction on unit_repairman for nearby player units applies to start point as well as current position.
+## 1.4a
+* Added unit_repairman helper to schedule periodic respawns of named units
 ## v1.3.1a
 * Added ap_utils.makeRocIfNameContains
 * Added overflow paging for comms menus
-* Added unit_repairman helper to schedule periodic respawns of named units
 ## v1.3a
 * Combined all scripts into one file
 * Fixed ap_utils.generateGroups issue with airbourne spawns
 * Added optional predicate for groups going idle in constant_pressure_set
 * Added ap_utils.getClosestLateralPlayer to help build predicates based on distance between player and units of a group
 * Idle check now checks all units in the group
-
 ## v1.2a
 * Respawnable_on_call.add should be called respawnable_on_call.new -- Fixed
 * Added respawnable_on_call:delete